dash board

+++ database/migrations/2024_12_25_035207_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string("name")->nullable();
             $table->integer("category_id")->nullable();
             $table->string("brand")->nullable();
-            $table->integer("price")->nullable();
+            $table->decimal("price", 10, 2)->nullable();
             $table->integer("discount")->default(0)->nullable();
             $table->integer("sell_price")->nullable();
             $table->integer("stock")->nullable();
